add org-checklist to automatically clear checkboxes on task state toggling
When moving a repeating task from TODO to DONE, the checkboxes are left as-is.
This means when the task repeats, it will show the checkboxes as checked. I want
repeating tasks to have their checkboxes cleared automatically. If
:RESET_CHECK_BOXES: t is set on the properties of the task, it will uncheck all
checked checkboxes in the task when the TODO state changes. This even applies
for repeating tasks, which still show as TODO when toggled due to how it's
represented.

org-checklist appears to come from org-plus-contrib, which is already installed.
The solution is to merely include it in the org-modules list.