hack star visibility back in
These atrocities are committed in the name of
https://github.com/hlissner/doom-emacs/issues/3076 in order to get stars
visible again and other things that can speed up org-mode per the Doom
Emacs readme. Some of these workarounds might be unnecessary but they
seem to be working currently and reducing it further has the effect of
hiding the stars again - so I will leave it as-is for now.