hand shell-mode indentation
Some of this is already done, but it's important to note that shell-mode
guesses indentation based on the existing document, which has lead me
here on more than one occasion.

The indentation around the =then= keyword is inconsistent and needlessly
wasteful of valuable horizontal space. This huge block includes a small
part that fixes it.